Convert uniqueness to Name, not Query
The original version of this used query as uniqueness constraint. Things were indexed by query. This proved hard to scale to supporting snapshot and non-snapshots. (which need identical queries)

This removes the entire `queries` hash, and moves all the rollup to the `confdb` one, and moves the uniqueness over to the name. 

This has some odd effects, but I think it's a bit cleaner